refactor: Update import paths for built-in styles
This commit updates the import paths for the built-in styles in the `built_in` directory. The import paths have been changed to reflect the new directory structure in the `core/style` directory. This change improves code organization and ensures that the correct styles are imported.
