+++ lib/src/core/style/factory/flat_colored_style.dart
