+++ lib/src/core/style/factory/minimal_style.dart
