+++ lib/src/core/style/factory/simple_style.dart
