+++ lib/src/widget/built_in/filled/filled.dart
@@ -3,7 +3,7 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:toastification/src/core/toastification_type.dart';
 import 'package:toastification/src/widget/built_in/built_in.dart';
-import 'package:toastification/src/widget/built_in/filled/filled_style.dart';
+import 'package:toastification/src/core/style/factory/filled_style.dart';
 import 'package:toastification/src/widget/built_in/widget/close_button.dart';
 
 class FilledToastWidget extends StatelessWidget {
