+++ lib/src/widget/built_in/flat_colored/flat_colored.dart
@@ -3,7 +3,7 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:toastification/src/core/toastification_type.dart';
 import 'package:toastification/src/widget/built_in/built_in.dart';
-import 'package:toastification/src/widget/built_in/flat_colored/flat_colored_style.dart';
+import 'package:toastification/src/core/style/factory/flat_colored_style.dart';
 import 'package:toastification/src/widget/built_in/widget/close_button.dart';
 
 class FlatColoredToastWidget extends StatelessWidget {
