+++ lib/toastification.dart
@@ -17,15 +17,15 @@ export 'src/widget/toast_animation.dart';
 
 // built-in widgets
 export 'src/widget/built_in/built_in.dart';
-export 'src/widget/built_in/built_in_style.dart';
+export 'src/core/style/built_in_style.dart';
 export 'src/widget/built_in/built_in_builder.dart' hide BuiltInBuilder;
 export 'src/widget/built_in/minimal/minimal.dart';
-export 'src/widget/built_in/minimal/minimal_style.dart';
+export 'src/core/style/factory/minimal_style.dart';
 export 'src/widget/built_in/filled/filled.dart';
-export 'src/widget/built_in/filled/filled_style.dart';
+export 'src/core/style/factory/filled_style.dart';
 export 'src/widget/built_in/flat/flat.dart';
-export 'src/widget/built_in/flat/flat_style.dart';
+export 'src/core/style/factory/flat_style.dart';
 export 'src/widget/built_in/simple/simple.dart';
-export 'src/widget/built_in/simple/simple_style.dart';
+export 'src/core/style/factory/simple_style.dart';
 export 'src/widget/built_in/flat_colored/flat_colored.dart';
-export 'src/widget/built_in/flat_colored/flat_colored_style.dart';
+export 'src/core/style/factory/flat_colored_style.dart';
